rename UserDto to UserResponseDto and add it in the dto/auth folder
